q4/web/requirements.txt
Django==2.2.1
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
mysqlclient==1.4.2
django-mysql==3.1.0
